requirements.txt
openai
langchain
streamlit
python-dotenv
langchain_openai
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu